backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
Markdown==3.4.1
mysql-connector-python==8.0.30
Pillow==9.2.0
protobuf==3.20.1
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
whitenoise==6.2.0
